DelfiAS.feature (Cucumber stuff) 2
http://samples.openweathermap.org/data/2.5/weather?q=London,uk&appid=b1b15e88fa797225412429c1c50c122a1

test

+++ src/test/java/delfiStepDefs/model/Weather.java
@@ -5,32 +5,31 @@ import java.math.BigDecimal;
 
 public class Weather {
 
-    private BigDecimal id;
+    private Integer id;
     private String main;
     private String description;
     private String icon;
 
 
-    public BigDecimal getId() {
+    public Integer getId() {
         return id;
     }
-
-    public void setId(BigDecimal id) {
+    public void setId(Integer id) {
         this.id = id;
     }
+
 
     public String getMain() {
         return main;
     }
-
     public void setMain(String main) {
         this.main = main;
     }
 
+
     public String getDescription() {
         return description;
     }
-
     public void setDescription(String description) {
         this.description = description;
     }
@@ -38,7 +37,6 @@ public class Weather {
     public String getIcon() {
         return icon;
     }
-
     public void setIcon(String icon) {
         this.icon = icon;
     }
